pyproject.toml
[project]
name = "wizard"
version = "2.2.6"
description = "Source-agnostic local memory layer for AI agents"
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "faker>=24.0",
    "fastmcp[tasks]>=3.2.4",
    "httpx>=0.27",
    "litellm>=1.83.7",
    "authlib>=1.6.11",
    "mako>=1.3.11",
    "python-multipart>=0.0.26",
    "notion-client>=3.0",
    "phonenumbers>=8.13",
    "pydantic-settings>=2.0",
    "sentry-sdk>=2.58.0",
    "streamlit>=1.35",
    "pandas>=2.0",
    "sqlmodel>=0.0.38",
    "tomli-w>=1.0.0",
    "typer>=0.12.0",
]

[project.scripts]
wizard = "wizard.cli.main:app"
wizard-server = "wizard.cli.serve:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wizard"]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.23.1",
    "ruff>=0.11.0",
    "wizard",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = ["integration: live API tests (require real credentials)"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "F",     # pyflakes (includes F401 unused imports)
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "C90",   # mccabe complexity (C901)
    "I",     # isort
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "INP",   # flake8-no-pep420 (INP001 — missing __init__.py)
]
ignore = [
    "B008",  # Depends() in argument defaults is the standard FastMCP/FastAPI pattern
]

[tool.ruff.lint.per-file-ignores]
"src/wizard/prompts.py" = [
    "E501",   # prompt templates contain long prose lines — wrapping harms readability
]
"src/wizard/alembic/**" = [
    "E501",   # auto-generated migration files contain long lines — do not reformat
    "W291",   # auto-generated migration files may have trailing whitespace
    "I001",   # auto-generated migration files have non-standard import ordering
]
"tests/**" = [
    "E402",   # module-level imports not at top — test files use late imports by design
    "E501",   # long lines in tests — assertion messages, mock data, URLs are clearer unsplit
    "SIM117", # nested with statements — test mocking often requires separate with blocks
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["wizard"]

[tool.uv.sources]
wizard = { workspace = true }
